Expose a simple API that combines Nathan's and my work
